enum 타입을 Types.swift 로 이동
